+++ src/test/carly/op.clj
@@ -24,22 +24,21 @@
   "Apply a sequence of operations to a system, returning a vector of pairs of
   the operations with their results."
   [system ops]
-  (mapv #(assoc % ::result (apply-op % system)) ops))
+  (mapv
+    (fn [op]
+      (assoc
+        op ::result
+        (try
+          (apply-op op system)
+          (catch Throwable t
+            t))))
+    ops))
 
 
 (defn run-ops!
   "Applies a sequence of operations in a separate thread. Returns a promise for
   the results of the application."
   [latch system thread-id ops]
-  #_
-  (let [results (promise)]
-    (doto (Thread.
-            (fn apply-ops-thread []
-              @latch
-              (deliver results (apply-ops! system ops)))
-            (format "test.carly/%s/%d" (Integer/toHexString (hash system)) thread-id))
-      (.start))
-    results)
   (future @latch (apply-ops! system ops)))