+++ src/test/carly/world.clj
@@ -6,10 +6,10 @@
     [test.carly.op :as op]))
 
 
-;; A world represents a point in time along a possible history. The model
-;; holds the current representation of the system, the history is the sequence
-;; of operations which have already happened, and pending is a map from thread
-;; ids to lists of pending operations for each thread.
+;; A world represents a point in time along a possible history. The `model`
+;; holds the current representation of the system, the `history` is the
+;; sequence of operations which have already happened, and `pending` is a map
+;; from thread ids to lists of pending operations for each thread.
 (defrecord World
   [model history pending futures])
 
@@ -96,13 +96,15 @@
 
 (defn run-linear
   "Steps a world forward to completion along a linear track. Returns a valid
-  terminal world if the operations end in a valid state, otherwise nil."
-  [world]
+  terminal world if the operations end in a valid state, otherwise nil. Calls
+  `f` with each world visited."
+  [world f]
   (when world
+    (f world)
     (if (end-of-line? world)
       ; Made it to the end of the world line with consistent results.
       world
       ; Step world forward. A nil here means the next operation result
       ; is invalid, so the observed worldline is inconsistent with the
       ; model.
-      (recur (step world)))))
+      (recur (step world) f))))