+++ CHANGELOG.md
@@ -9,6 +9,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ...
 
+## [0.3.0] - 2017-06-09
+
+### Changed
+- `:context-gen` option shortened to `:context`.
+- Exceptions thrown during an operation are now caught and used as the operation
+  result.
+- World worker thread poll time shortened to 100 ms.
+
+### Fixed
+- World states visited in a linear scan are properly counted.
+
 ## [0.2.0] - 2017-05-25
 
 This release includes substantial optimizations to searching for valid
@@ -27,6 +38,7 @@ worlds (those with fewer possible futures) searched first.
 
 Initial project release.
 
-[Unreleased]: https://github.com/greglook/test.carly/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/greglook/test.carly/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/greglook/test.carly/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/greglook/test.carly/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/greglook/test.carly/tag/0.1.0